backend/requirements.txt
Flask
flask-cors
torch
torchvision
tensorflow
Pillow
numpy